requirements.txt
torch>=2.0.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
